requirements.txt
fastapi
uvicorn
litellm
boto3
python-dotenv
pydantic
